pyproject.toml
[tool.poetry]
name = "pythonweb-hw012"
version = "0.2.0"
description = "REST API for managing contacts"
authors = ["Dmytro Zhuk <dm.zhuk@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
fastapi-limiter = "^0.1.6"
fastapi-mail = "^1.5.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.0"
pydantic = { version = "^2.0.0", extras = ["email"] }
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"
python-dateutil = "^2.9.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
aiosmtplib = "^3.0.1"
cloudinary = "^1.41.0"
asyncpg = "^0.30.0"
libgravatar = "^1.0.4"
redis = "^6.2.0"
alembic = "^1.16.2"
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"  # Updated to match your main dependencies
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
sphinx = "^8.2.3"
aiosqlite = "^0.21.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"
 
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"